Reverse argument order in R.tap. Solve #453
R.tap took its arguments value first, function second but it was
generally agreed that the function is more extensible if the arguments
were reversed.

The type check of the first argument to R.tap was removed which makes
the function behave like other functions in ramda where the user is
expected to pass in the correct arguments.