Adding functions pCompose and pPipe.
R.pCompose works like R.compose, but handles promises.
The resulting function will evaluate to a promise if
any one of the composed functions returns a promise.
If all of the composed functions return something other
than a promise, then the behaviour is identical to that
of R.compose(). In this sense, R.pCompose() is backwards
compatible with R.compose.

R.pPipe is like R.pCompose but takes the arguments in the opposite order.