Add internal function _depricationWarning.
The _depricationWarning function provides an easy way to show a
deprication message around a function that has been renamed. By default
it prints the name of the old function and what it has been renamed to.
It also provides the possibility to add an additional message.

_depricationWarning makes sure that it only prints the error message
once per depricated function.