Remove _symTransformer until official spec is worked out.
Prompted by cognitect-labs/transducers-js#20 from @tgriesser, removing
the internal transformer symbol until issues are worked out. This symbol
is currently only supported in jlongster/transducers.js and
interoperability is questionable as (Symbol('transformer') !==
Symbol('transformer')).

This feature was not documented, nor do the tests depend on it, and
only effects an undocumented feature of into. Better to remove it now
until details are worked out.  If an "official spec" is agreed upon,
I will open a new PR.