Add through function
This function takes a list of functions and an object. It returns the
first defined value from the list of functions. It returns undefined if
nothing is found.

The `through` function is intended to build a disjunction on the values
of `f`, `g`, `h`:

    f' = f(a) || g(a) || h(a)

Where the disjunction is short-circuited by a any value other than
`undefined`.