depthReduce, a stoppable depth-first reduce for deep iterables
Returns a single item by running a depth-first iteration through a given
multi-dimensional iterable (list, object, nested objects and arrays or
cyclic object), successively calling the iterator function and passing
it an accumulator value, the current value from the iterable, the key
where the current value is stored in the iterable, a list with the keys
of the parent iterables of the current position, and then passing the
result to the next call.

The iterator function receives four values: *(acc, value, key,
parentsKeys)*.  It's expected that the iterator returns it's results
wrapped in an array, of which only the first position is considered. If
the iterator returns an empty array, depthReduce immediately stops,
returning the last non-empty result.

See: https://github.com/ramda/ramda/issues/2128#issuecomment-287766254