reduce: dispatch to `reduce` method of list as a last resort.
When using transducers, the reduce implementation is required
to support the shortcutting feature (see `reduced`). This
feature cannot be guaranteed by the `reduce` method of any
general implementation, so it is better to use it as last
resort (giving precedence to reducing with iterable/iterator,
where the full control over is retained).