Reimplement prop via path
As discussed in #2059 we are unsatisfied that `prop` and `path` have
different behaviour when passing `undefined` as the object.

We could not yet agree which behaviour we want to have but at least
we want to have the same behaviour. To achieve that, we now
reimplement `prop` with `path`.

So if we ever define the behaviour of `path` then `prop`
automatically also has the same behaviour.