Undo change to mapAccumRight accumulator argument order
The order of arguments expected by mapAccumRight's accumulator function
was reversed in #1975 (c0292498fa798118b70c8d6298f40c74211e1225), to be
consistent with reduceRight. There was a bug in the implementation and
it was decided to undo the change and let the argument order be
consistent with Haskell's mapAccumR in #2386.