Add size snapshot rollup plugin
This diff adds size snapshot plugin for rollup which on every build
writes snapshot of umd with actual, minified and gzipped sizes. This
allows to track changes in size on every commit.

Also this diff combines development and production rollup configs in one
multiconfig to run node with rollup only once.