Add file extensions to imports and exports in `source/`
This enables a developer to load Ramda via import statements directly in the
browser without bundling/transpilation.

Before, imports in the Ramda source only included the file name without
extension, eg. `import curryN from './curryN'`. Most bundlers will
assume a default extension and find the file regardless. In the browser,
the exact file path that was specified in the import statement will be
requested, and the file can usually only be served by the webserver or
located on the file system if the file path matches exactly. Thus, the
complete file path must be used in import statements to be able to load
Ramda via import statements without some kind of bundling/transpilation
process.

Fixes #2587.