fix: rename then to andThen (#2772)
BREAKING CHANGE

The addition of R.then in Ramda 0.26 made the R object
a Promise-like object, and meant that Promise.resolve() will attempt
to resolve it by calling the then() method. R is not a Promise, and
to reassure JavaScript of this fact, R.then is renamed to R.andThen.