Fix traverse type signature in documentation comment
Since `traverse = sequence of (map f traversable)`, and
`sequence :: (x -> f x) -> t (f x) -> f (t x)`,
`f :: a -> f b` and `traversable :: t a`, the type of `(map f traversable)`
is `t (f b)`. Therefore the type parameter `x` in `sequence` is `b`,
and the type of the `of` parameter needs to be `b -> f b`.