Remove placeholder check from identical
This fixes [#2533](https://github.com/ramda/ramda/issues/2533), were using identical with cross-origin browser window objects throws a DOMException.

The reason for the exception is the checking for the placeholder property in the `curry*` functions.

Since placeholder support makes no sense for commutative functions it can be removed from identical. This is done here by rewriting `identical` as `isObject` in a self-written `curry2` wrapper.