Restore functions removed after v0.27.0 to avoid breaking changes in v0.28.0
This partially reverts commit 3366df607287bf5cf3b9f870ad5b463198f334e5.

This commit temporarily restores the functions removed
after the initial v0.27 release so as to avoid causing
breaking changes in the next minor release v0.28.0

This commit also temporarily restore a dependency to `jsverify`
needed for the tests that have also been restored as a
result of reverting the commit.